give act_sequence its own module, so act-tools fits under the backstop
#423 was written against `release/v2.12.0`, where `act-tools.ts` had room. On
`main`, 48 commits further along, the same addition takes it to 1003 lines and
`max-lines` refuses it. The cap is the backstop for cohesion, so this wants a
seam, not a bigger number.

`reticle_act_sequence` is the seam. Its two neighbours drive ONE action and
grade its consequence; this one compiles a list of steps and runs them inside a
single action window, reporting per step. Different unit of work.

The first attempt moved the shared `actCommand` helper out instead, and
`dispatch-attribution.test.ts` caught it: that separates an ACT dispatch from
the `beginAction` window that must wrap it, and an unattributed effect is
classified as ambient churn, which teaches the settle oracle to ignore the very
regions the app reacts in. Adding the new file to that guard's escape hatch
would have been the wrong repair — the hatch is for non-dispatching references,
and the helper really did dispatch. This seam keeps every dispatch in the same
file as its attribution.

No behaviour change: the tool is unmoved, re-homed and re-registered in the same
array position.
