+++ packages/server/src/tools/act-tools.ts
@@ -1,3 +1,6 @@
+import { normalizeQueryArgs } from './query-shape.js';
+import type { Session } from '../session/session.js';
+import { resolveTargetRef, type TargetResolution } from './resolve-target.js';
 /**
  * Action tools — reticle_act, reticle_act_sequence, reticle_act_and_wait. Split out of tools.ts to keep
  * that file under the line cap and assembled back into the tool list there via ...ACT_TOOLS; the
@@ -5,15 +8,16 @@
  */
 import { z } from 'zod';
 import { aliasParam } from './alias-args.js';
-import { actCommand, resolveActTarget, actionTypeEnum, ACTION_TYPE_LIST } from './act-plumbing.js';
+import { ACT_SEQUENCE_TOOL } from './act-sequence-tool.js';
 import { timeoutMsSchema } from './numeric-bounds.js';
-import { captureAct, compileSequenceStep } from '../flows/replay.js';
+import { captureAct } from '../flows/replay.js';
 import {
   ActionType,
   ActionWarning,
   CaptureLoss,
   DEFAULT_ASSERT_TIMEOUT_MS,
   InputMode,
+  ReticleCommand,
   Verified,
   VerifiedReason,
   PredicateKind,
@@ -74,12 +78,105 @@ import {
   PAUSED_NO_VERDICT,
 } from '../session/control-envelope.js';
 import { asString, asNumber, asRecord, sourceOf } from './tools-helpers.js';
-import { describeStepResult, runStepWithStaleRetry } from './act-sequence-retry.js';
-import { assertSequenceSteps, dispatchAct, preflightAct } from './act-preflight.js';
+import { dispatchAct, preflightAct } from './act-preflight.js';
 import { followLostObservation } from './act-observation.js';
 import { type ToolDef, intentArg, sessionIdShape } from './tool-kit.js';
 import { asActionType, gradeOf } from './act-helpers.js';
-import { tryRealInput } from './real-input-attempt.js';
+import { tryRealInput, rewriteUploadArgs } from './real-input-attempt.js';
+
+/**
+ * Single dispatch point for every ACT and ACT_SEQUENCE command.
+ *
+ * This is the seam the reviewer asked for: instead of wiring rewriteUploadArgs at three separate
+ * call sites (reticle_act, reticle_act_and_wait, reticle_act_sequence) we intercept once here,
+ * which also covers flow-replay, crawl, and any future dispatch site that goes through this helper.
+ *
+ * captureAct is called AFTER this, so the flow-recording captures the pre-rewrite args (the path
+ * the agent actually wrote) rather than the base64 blob — replay would otherwise send 750 KiB of
+ * base64 to the browser as if it were an inline content call.
+ */
+export async function actCommand(
+  deps: Parameters<typeof rewriteUploadArgs>[0],
+  session: {
+    command: (
+      name: string,
+      args: Record<string, unknown>,
+      timeout?: number,
+    ) => Promise<import('@reticlehq/core').CommandResult>;
+  },
+  actArgs: Record<string, unknown>,
+  timeoutMs?: number,
+): Promise<import('@reticlehq/core').CommandResult> {
+  const rewritten = await rewriteUploadArgs(
+    deps,
+    'string' === typeof actArgs['action'] ? actArgs['action'] : '',
+    asRecord(actArgs['args']),
+  );
+  const bridgeArgs: Record<string, unknown> = { ...actArgs, args: rewritten };
+  return timeoutMs !== undefined
+    ? session.command(ReticleCommand.ACT, bridgeArgs, timeoutMs)
+    : session.command(ReticleCommand.ACT, bridgeArgs);
+}
+
+/**
+ * Narrow the wire's `action` to a real ActionType, or undefined.
+ *
+ * It used to be `asString(args['action']) ?? ''`, so an unknown or missing action became the empty
+ * string and travelled on — reaching the browser as a command it could not perform, and reported back
+ * as a generic failure rather than "that is not an action". Validate at the boundary, per the project's
+ * unknown-plus-narrowing rule, so a typo is rejected where it can still be explained.
+ */
+
+/**
+ * The action vocabulary, derived from ActionType — never retyped.
+ *
+ * The description used to list thirteen actions while ActionType had seventeen: blur, upload, drag
+ * and webmcp were real, callable, and undocumented, because a hand-copied list drifts the moment
+ * someone adds an arm. Deriving both the schema and the prose from the enum makes that impossible.
+ *
+ * The handler already refused an unknown action with a good message; the schema now refuses it
+ * one layer earlier, before any session is resolved or any work is done.
+ */
+const ACTION_TYPE_VALUES = Object.values(ActionType);
+const ACTION_TYPE_LIST = ACTION_TYPE_VALUES.join(' | ');
+export const actionTypeEnum = z.enum(ACTION_TYPE_VALUES as [string, ...string[]]);
+
+/**
+ * Resolve an action's element: an explicit `ref`, or a `target` query resolved in the SAME call.
+ *
+ * Requiring a ref meant every verification paid a `reticle_query` turn first just to learn one
+ * string, and the advertised tool surface is re-sent on every turn — measured on the wire, a
+ * two-turn verification spent 10,756 of 11,235 tokens on schema and 479 on the actual answers. The
+ * lookup still happens; it just stops costing a round trip through the model.
+ *
+ * `ref` wins when both are given, because it is the more specific instruction and silently
+ * preferring the query would act on something the caller did not name.
+ */
+export async function resolveActTarget(
+  session: Session,
+  args: Record<string, unknown>,
+): Promise<TargetResolution> {
+  const ref = asString(args['ref']);
+  if (ref !== undefined && ref.length > 0) return { kind: 'ref', ref };
+  const target = args['target'];
+  if (target === undefined) {
+    return {
+      kind: 'error',
+      message:
+        'pass `ref` (from reticle_query/reticle_snapshot) or `target` (e.g. { testid } or { role, name }).',
+    };
+  }
+  const q = normalizeQueryArgs(asRecord(target));
+  const out = await session.command(ReticleCommand.QUERY, {
+    by: q['by'],
+    value: q['value'],
+    name: q['name'],
+    scope: q['scope'],
+  });
+  if (!out.ok) return { kind: 'error', message: out.error ?? 'target query failed' };
+  const elements = asRecord(out.result)['elements'];
+  return resolveTargetRef(Array.isArray(elements) ? elements : []);
+}
 
 export const ACT_TOOLS: ToolDef[] = [
   {
@@ -225,125 +322,6 @@ export const ACT_TOOLS: ToolDef[] = [
           settledOutcome,
           true === settledOutcome ? session.elapsed() - since : undefined,
         );
-      }
-    },
-  },
-  {
-    name: ReticleTool.ACT_SEQUENCE,
-    // The example is required for a core tool, and this one carries weight: the measured loop it
-    // replaces is literally a login form driven as three separate reticle_act calls (98 clicks and
-    // 21 fills inside looping sessions, 2026-08-10/11). Showing fill -> fill -> click is showing the
-    // exact shape an agent otherwise spends three round trips on.
-    example: {
-      steps: [
-        { ref: 'e12', action: 'fill', args: { value: 'a@b.com' } },
-        { ref: 'e13', action: 'fill', args: { value: 'hunter2' } },
-        { ref: 'e14', action: 'click' },
-      ],
-    },
-    description:
-      'Run multiple actions in order (fill -> fill -> submit) in ONE round-trip. Prefer this over repeating reticle_act for a multi-step journey, then assert its consequence once. Returns per-step effects[] (see reticle_act).',
-    inputSchema: {
-      steps: z
-        .array(z.record(z.unknown()))
-        .describe(
-          'Ordered list of { ref, action, args? } objects. Each step is equivalent to one reticle_act call; put confirmDangerous:true in a destructive step args object.',
-        ),
-      timeout_ms: timeoutMsSchema
-        .optional()
-        .describe(
-          'Per-step timeout in milliseconds. Default: 8000. Each step gets this budget independently.',
-        ),
-      ...sessionIdShape,
-    },
-    outputSchema: {
-      since: z.number(),
-      dispatched: z.boolean(),
-      completed: z.number(),
-      stalled_at: z.number().optional(),
-      steps: z.array(z.record(z.unknown())).optional(),
-      result: z.unknown().optional(),
-      session: z
-        .object({ lastSeenMs: z.number(), throttled: z.boolean(), focused: z.boolean() })
-        .optional(),
-      // Short-circuits to pausedShortCircuit while paused — declare its fields (drained-once guidance).
-      ...pausedOutputShape,
-    },
-    handler: async (deps, args) => {
-      const session = deps.sessions.resolve(asString(args['sessionId']));
-      const paused = pausedShortCircuit(session);
-      if (paused !== undefined) return paused;
-      const since = session.elapsed();
-      session.beginAction(ReticleTool.ACT_SEQUENCE, asRecord(args));
-      try {
-        const inputSteps = Array.isArray(args['steps']) ? args['steps'] : [];
-        assertSequenceSteps(inputSteps);
-        const perStepTimeout = 'number' === typeof args['timeout_ms'] ? args['timeout_ms'] : 8000;
-        const stepResults: Record<string, unknown>[] = [];
-        let stalledAt: number | undefined;
-
-        // DIVERGENCE: live sends N individual ACT commands (for per-step timeout + progress);
-        // replay sends one batched ACT_SEQUENCE command (flows/replay.ts:294). A bug in either is
-        // invisible from the other — cover both when changing sequence semantics.
-        for (let i = 0; i < inputSteps.length; i++) {
-          const step = asRecord(inputSteps[i]);
-          try {
-            // One retry when the ref went stale under a re-render — see act-sequence-retry.ts.
-            const outcome = await runStepWithStaleRetry(
-              () =>
-                actCommand(
-                  deps,
-                  session,
-                  { ref: step['ref'], action: step['action'], args: step['args'] ?? {} },
-                  perStepTimeout,
-                ),
-              session,
-              since,
-              perStepTimeout,
-            );
-            if (!outcome.ok) {
-              stalledAt = i;
-              stepResults.push({
-                ref: step['ref'],
-                action: step['action'],
-                dispatched: false,
-                error: outcome.error ?? 'step failed',
-              });
-              break;
-            }
-            stepResults.push(describeStepResult(step, asRecord(outcome.result)));
-          } catch (err: unknown) {
-            stalledAt = i;
-            stepResults.push({
-              ref: step['ref'],
-              action: step['action'],
-              dispatched: null,
-              timedOut: true,
-              error: err instanceof Error ? err.message : 'step timed out',
-            });
-            break;
-          }
-        }
-
-        const completed = stalledAt ?? inputSteps.length;
-        if (completed > 0) {
-          session.lastAct.markActed(since, undefined, undefined);
-        }
-        if (deps.recordings.active().length > 0 && stalledAt === undefined) {
-          deps.recordings.capture(
-            compileSequenceStep(args, { count: inputSteps.length, steps: stepResults }),
-          );
-        }
-        return withControl(session, {
-          since,
-          dispatched: completed > 0,
-          completed,
-          ...(stalledAt !== undefined ? { stalled_at: stalledAt } : {}),
-          steps: stepResults,
-          ...healthEnvelope(session),
-        });
-      } finally {
-        session.finishAction();
       }
     },
   },
@@ -903,4 +881,6 @@ export const ACT_TOOLS: ToolDef[] = [
       }
     },
   },
+  // Split into its own module for size; still shipped as one of the acting tools.
+  ACT_SEQUENCE_TOOL,
 ];
